+++ server/api/announcement/index.spec.js
@@ -5,11 +5,24 @@
 var proxyquire = require('proxyquire').noPreserveCache();
 
 var announcementCtrlStub = {
-  index: 'announcementCtrl.index'
+  index: 'announcementCtrl.index',
+  upsert: 'announcementCtrl.upsert',
+  destroy: 'announcementCtrl.destroy'
+};
+
+var authServiceStub = {
+  isAuthenticated() {
+    return 'authService.isAuthenticated';
+  },
+  hasRole(role) {
+    return `authService.hasRole.${role}`;
+  }
 };
 
 var routerStub = {
-  get: sinon.spy()
+  get: sinon.spy(),
+  put: sinon.spy(),
+  delete: sinon.spy()
 };
 
 // require the index with our stubbed out modules
@@ -19,7 +32,8 @@ var announcementIndex = proxyquire('./index.js', {
       return routerStub;
     }
   },
-  './announcement.controller': announcementCtrlStub
+  './announcement.controller': announcementCtrlStub,
+  '../../auth/auth.service': authServiceStub
 });
 
 describe('Announcement API Router:', function() {
@@ -32,6 +46,24 @@ describe('Announcement API Router:', function() {
       expect(routerStub.get
         .withArgs('/', 'announcementCtrl.index')
         ).to.have.been.calledOnce;
+    });
+  });
+
+  describe('PUT /api/announcement/:id', function() {
+    it('should be authenticated and route to announcement.controller.upsert', function(done) {
+      expect(routerStub.put
+        .withArgs('/:id', 'authService.hasRole.admin', 'announcementCtrl.upsert')
+        ).to.have.been.calledOnce;
+      done();
+    });
+  });
+
+  describe('DELETE /api/announcement/:id', function() {
+    it('should verify admin role and route to announcement.controller.destroy', function(done) {
+      expect(routerStub.delete
+        .withArgs('/:id', 'authService.hasRole.admin', 'announcementCtrl.destroy')
+        ).to.have.been.calledOnce;
+      done();
     });
   });
 });